% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_mstte_fit.R
\name{as.data.frame.posterior_mstte_fit.stanmstte}
\alias{as.data.frame.posterior_mstte_fit.stanmstte}
\title{Method for stanmstte}
\usage{
\method{as.data.frame}{posterior_mstte_fit.stanmstte}(x, ...)
}
\description{
Method for stanmstte
}
