% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_y}
\alias{get_y}
\alias{get_x}
\title{Extract X, Y or Z from a stanreg object}
\usage{
get_y(object, ...)

get_x(object, ...)
}
\arguments{
\item{...}{Other arguments passed to methods. For a \code{stanmvreg} object
this can be an integer \code{m} specifying the submodel.}
}
\value{
For \code{get_x} and \code{get_z}, a matrix. For \code{get_y}, either
  a vector or a matrix, depending on how the response variable was specified.
}
\description{
Extract X, Y or Z from a stanreg object
}
\keyword{internal}
