% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{summary.stanidm}
\alias{summary.stanidm}
\alias{print.summary.stanidm}
\alias{as.data.frame.summary.stanidm}
\title{Summary method for stanidm objects}
\usage{
\method{summary}{stanidm}(object, pars = NULL, regex_pars = NULL,
  probs = NULL, ..., digits = 1)

\method{print}{summary.idm}(x, digits = max(2, attr(x, "print.digits")),
  ...)

\method{as.data.frame}{summary.stanidm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.stanidm"}.}
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics
(Monte Carlo error, effective sample size, Rhat).
}
