% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_mat.R
\name{trans_mat}
\alias{trans_mat}
\title{Create a transition matrix
--------------------------}
\usage{
trans_mat(x, names)
}
\arguments{
\item{x:}{List of possible transitions.
x[[i]] consists of a vector of state numbers
reachable from state i.}

\item{names:}{Character vector of state names, having the same length
as x.}
}
\description{
In a competing risk setting state 2 is reachable from state 1.
State 3 is absorbing, reachable from states 1 and 2.
}
\examples{
\donttest{
   # For a competing risks setting
   trans_mat( x = list( c(2, 3), c(3), c() ) )
   # For consecutive states
   trans_mat( x = list(c(2), 3, c() ) )
   }
}
\references{
Liesbeth C. de Wreede, Marta Fiocco, Hein Putter
(2011). mstate: An R Package for the Analysis of
Competing Risks and Multi-State Models. Journal of
Statistical Software, 38(7), 1-30. URL
http://www.jstatsoft.org/v38/i07/.
}
